+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added: After clicking the "Login with Discord" button, a link and a QR Code are displayed instead of opening the default browser directly.
 - Changed: The relative time to start/finish in Async Race rooms now update as time passes, and the window reacts appropriately when the race starts and finishes.
+- Fixed: A server issue that caused async race exporting of AM2R and MSR to let you easily view the spoiler.
 - Removed: The initial Randovania screen no longer contains the list of supported games.
 
 ### AM2R